requirements.txt
sentence-transformers
faiss-cpu
transformers
gtts
deep-translator
SpeechRecognition
gradio
pandas
numpy
torch